requirements.txt
flask
pandas
numpy
tensorflow
pillow
scikit-learn
matplotlib